% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsqexpdistr.R
\name{tsqexpdistr}
\alias{tsqexpdistr}
\alias{dtsqexp}
\alias{ptsqexp}
\alias{qtsqexp}
\alias{rtsqexp}
\title{The q-exponential distribution}
\source{
\code{[dpq]tsqexp} are calculated directly from the definitions, \code{rtsqexp} uses inversion.
}
\usage{
dtsqexp(x, shape, rate = 1, log = FALSE)

ptsqexp(q, shape, rate = 1, lower.tail = TRUE, log.p = FALSE)

qtsqexp(p, shape, rate = 1)

rtsqexp(n, shape, rate = 1)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{shape, rate}{shape and rate parameters, the latter defaulting to 1.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \emph{P[X ≤ x]}, otherwise, \emph{P[X > x]}.}

\item{p}{vector of probabilities}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}
}
\value{
\code{dtsqexp} gives the density, \code{ptsqexp} gives the distribution function, \code{qtsexp} gives the quantile function, and \code{rtsqexp} generates random deviates.

The length of the result is determined by \code{n} for \code{rtsqexp}, and is the maximum of the lengths of the numerical arguments for the other functions.

The numerical arguments other than \code{n} are recycled to the length of the result. Only the first elements of the logical arguments are used.
}
\description{
Density, distribution function, quantile functions and random generation for the \emph{q}-exponential distribution with shape \code{shape} and rate \code{rate}.
}
\details{
If \code{shape} and \code{rate} are not specified, they assume the default value of 1.

If \code{shape} is equal to 1, the \emph{q}-exponential distribution becomes the exponential.
}
\note{
The cumulative hazard \emph{H(t) = -log(1 - F(t))} is \code{-ptsqexp(t, q, r, lower = FALSE, log = TRUE)}.
}
\examples{
x <- c(0, rlnorm(50))
all.equal(dtsqexp(x, shape = 1), dexp(x))
all.equal(ptsqexp(x, shape = 1, rate = 1/pi), pexp(x, rate = 1/pi))
all.equal(qtsqexp(x/11, shape = 1, rate = 1/pi), qexp(x/11, rate = 1/pi))
}
\references{
Tsallis, C. Nonadditive entropy and nonextensive statistical mechanics-an overview after 20 years. Braz. J. Phys. 2009, 39, 337–356.
Picoli, S. Jr.; Mendes, R. S.; Malacarne, L. C. (2003). "q-exponential, Weibull, and q-Weibull distributions: an empirical analysis". Physica A: Statistical Mechanics and Its Applications. 324 (3): 678–688. arXiv:cond-mat/0301552. Bibcode:2003PhyA..324..678P. doi:10.1016/S0378-4371(03)00071-2. S2CID 119361445.
}
\seealso{
\code{\link[=tsqlog]{tsqlog()}} for the \emph{q}-logarithm and \emph{q}-exponential
}
