% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsqgauss.R
\name{tsqgaussdistr}
\alias{tsqgaussdistr}
\alias{dtsqgauss}
\alias{rtsqgauss}
\title{The q-Gaussian distribution}
\source{
\code{[dtsqgauss} is calculated directly from the definitions, \code{rtsqgauss} uses inversion.
}
\usage{
dtsqgauss(x, shape, location = 0, scale = 1, log = FALSE)

rtsqgauss(n, shape, location = 0, scale = 1)
}
\arguments{
\item{x}{vector of quantiles}

\item{shape}{shape parameter}

\item{location}{location parameter}

\item{scale}{scale parameter}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}
}
\value{
\code{dtsqgauss} gives the density, and \code{rtsqgauss} generates random deviates.

The length of the result is determined by \code{n} for \code{rtsqgauss}, and is the maximum of the lengths of the numerical arguments for the other functions.

The numerical arguments other than \code{n} are recycled to the length of the result. Only the first elements of the logical arguments are used.
}
\description{
Density, distribution function, quantile functions and random generation for the \emph{q}-Gaussian distribution with shape \code{shape} and scale \code{scale}.
}
\details{
If \code{shape1} is equal to 1, the \emph{q}-Gaussian distribution becomes the normal distribution.
}
\examples{
x <- c(0, rlnorm(50))
all.equal(dtsqgauss(x, shape = 1, scale = 2), dnorm(x, mean = 0, sd = 2))
}
\references{
Umarov, Sabir; Tsallis, Constantino; Steinberg, Stanly (2008). "On a q-Central Limit Theorem Consistent with Nonextensive Statistical Mechanics" (PDF). Milan J. Math. Birkhauser Verlag. 76: 307–328. doi:10.1007/s00032-008-0087-y. S2CID 55967725.
 W. Thistleton, J.A. Marsh, K. Nelson and C. Tsallis, Generalized Box–Muller method for generating q-Gaussian random deviates, IEEE Transactions on Information Theory 53, 4805 (2007)
}
\seealso{
\code{\link[=tsqlog]{tsqlog()}} for the \emph{q}-logarithm and \emph{q}-exponential
}
