% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsqlog.R
\name{tsqlog}
\alias{tsqlog}
\alias{tsqexp}
\title{q-logarithms and q-exponentials}
\usage{
tsqlog(x, q = 1)

tsqexp(x, q = 1)
}
\arguments{
\item{x:}{a numeric vector.}

\item{q:}{a \emph{q} parameter}
}
\value{
A vector of the same length as \code{x} containing the transformed values. \code{tsqlog} and \code{tsqexp} gives the same results as \code{log} and \code{exp} if \code{q} is 1.
}
\description{
\code{tsqlog} computes the q-logarithm function.

\code{tsqexp} computes the q-exponential function.
}
\examples{
tsqlog(tsqexp(3));
tsqlog(tqexp(5, q = 2), q = 2);
}
\references{
Tsallis, Constantino (1994). "What are the numbers that experiments provide?". Química Nova. 17: 468.
}
